+++ ark/repo/scratch.ts
@@ -5,3 +5,20 @@ const User = type({
 	platform: "'android' | 'ios'",
 	"version?": "number | string"
 })
+
+const selected = User.select({
+	kind: "domain",
+	where: d => d.domain === "string"
+})
+
+const ConfiguredUser = User.configure(
+	{ description: "A STRING" },
+	{
+		kind: "domain",
+		where: d => d.domain === "string"
+	}
+)
+
+ConfiguredUser.get("name").description // A STRING
+ConfiguredUser.get("platform").description // "android" | "ios"
+ConfiguredUser.get("version").description // a number, A STRING or undefined